moved PrecursorIonCrystal to VanillaPrefabPatching folder
